Media visualization program
This is a custom-written visualization program for displaying a `.csv` sheet of duration data. It's written in Java and runs using the Processing coding environment.

The input-csv should be formatted as
Date | Start Time | End Time | Entry Title
:---: | :---: | :---: | :---:
 mm/dd/yy | 0-1 decimal as starting time | 0-1 decimal as end time | TV Show Title
... | ... | ...

Each date is visualized as one column, with time progressing upward from 1 AM the previous day to 1 AM the next day. There are a few toggle-able visualization attributes; adjustable opacity for a specific TV show (i.e. look at all the times when `The Office` was watched), look at what time the show started at, and look at the top N shows in terms of how much time was spent watching them.

This can also work with non-Media data so long as the `isMedia` bool is changed to false, and the data is structured the same way.