Live music visualizer
This is a very work in progress project, currently it live-visualizes whatever song is both in the `/data/` folder, and is written inside `minim.loadFile("[SONG NAME].mp3", 512)`.

It visualizes the relative loudness at each segment of time (usually 3 - 5 seconds) in a clockwise fashion as a line projecting radially outward. 

I've borrowed a lot of this code from other projects, especially the implementation of the minim class.